// src/api/generated/models/createApiKeyData.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * ToggleFlow API
 * ToggleFlow is a Feature Flag Management Platform that allows teams to
 * manage feature flags, environments, experiments, API keys, segments,
 * webhooks, and analytics through a REST API.
 *
 * OpenAPI spec version: 1.0.0
 */

export interface CreateApiKeyData {
  key?: string;
  id?: string;
  /** @nullable */
  name?: string | null;
  prefix?: string;
  /** @nullable */
  expiresAt?: string | null;
  createdAt?: string;
  message?: string;
}